update book module

+++ app/src/main/java/me/stupidme/cooker/view/book/BookNowFragment.java
@@ -3,7 +3,7 @@ package me.stupidme.cooker.view.book;
 
 import android.os.Bundle;
 import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
+import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.helper.ItemTouchHelper;
 import android.util.Log;
@@ -11,7 +11,6 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
@@ -52,8 +51,10 @@ public class BookNowFragment extends BookBaseFragment implements BookDialog.Book
         View view = super.onCreateView(inflater, container, savedInstanceState);
         if (view != null) {
             mFab = (FloatingActionButton) view.findViewById(R.id.fab);
-
-            mFab.setOnClickListener(v -> mDialog.show());
+            mFab.setOnClickListener(v -> {
+                mDialog.reset();
+                mDialog.show();
+            });
         }
         return view;
     }
@@ -91,16 +92,31 @@ public class BookNowFragment extends BookBaseFragment implements BookDialog.Book
                         BookBean bean = mDataSet.get(position);
                         mDataSet.remove(position);
                         mAdapter.notifyItemRemoved(position);
+//
+//                        Snackbar.make(viewHolder.itemView,
+//                                getString(R.string.snackbar_text_cooker_fragment),
+//                                1000)
+//                                .setAction("DELETE", v -> {
+//                                    mPresenter.deleteBook(bean);
+//                                }).show();
+//
+//                        mDataSet.add(position, bean);
+//                        mAdapter.notifyItemInserted(position);
+//
 
-                        Snackbar.make(viewHolder.itemView,
-                                getString(R.string.snackbar_text_cooker_fragment),
-                                1000)
-                                .setAction("DELETE", v -> {
+                        new AlertDialog.Builder(viewHolder.itemView.getContext())
+                                .setMessage(getString(R.string.snackbar_text_cooker_fragment))
+                                .setTitle(getString(R.string.tips_title))
+                                .setNegativeButton("CANCEL", (dialog12, which) -> {
+                                    dialog12.dismiss();
+                                    mDataSet.add(position, bean);
+                                    mAdapter.notifyItemInserted(position);
+                                })
+                                .setPositiveButton("DELETE", (dialog1, which) -> {
+                                    dialog1.dismiss();
                                     mPresenter.deleteBook(bean);
-                                }).show();
-
-                        mDataSet.add(position, bean);
-                        mAdapter.notifyItemInserted(position);
+                                })
+                                .show();
                     }
                 };
         ItemTouchHelper helper = new ItemTouchHelper(callback);
@@ -131,14 +147,11 @@ public class BookNowFragment extends BookBaseFragment implements BookDialog.Book
     @Override
     public void onSave(Map<String, String> map) {
         Log.v("BookNowFragment", "onSave()");
+        mPresenter.insertBook(map);
     }
 
     @Override
     public List<String> getCookerNames() {
-        List<String> list = new ArrayList<>();
-        list.add("CookerA");
-        list.add("CookerB");
-        list.add("CookerC");
-        return list;
+        return mPresenter.queryCookerNamesFromDB();
     }
 }
